pyproject.toml
[project]
name = "tiktok-reup-webapp"
version = "1.1.0"
description = "A web application for re-uploading TikTok videos"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "moviepy>=2.2.0",
    "flask-sqlalchemy>=3.0.0",
    "bcrypt>=4.0.0",
    "ffmpeg-python>=0.2.0",
    "requests>=2.32.0",
    "curl-cffi>=0.12.0",
    "pyrogram>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]
